pyproject.toml
[tool.poetry]
name = "cooks-recipes"
version = "0.1.0"
description = ""
authors = ["Uliana Semyonova <i@ulianasemyonova.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
